CH8/requirements.txt
python-dotenv>=1.0.1,<2.0.0
langchain==0.3.0
langchain-core==0.3.0
langchain-google-genai<3.0.0
streamlit>=1.28.0
google-generativeai>=0.3.0